first push

+++ gnl/get_next_line.h
@@ -6,7 +6,7 @@
 /*   By: mtacunan <mtacunan@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/12/07 22:04:56 by mtacunan          #+#    #+#             */
-/*   Updated: 2021/12/09 16:48:12 by mtacunan         ###   ########.fr       */
+/*   Updated: 2022/04/07 12:05:37 by mtacunan         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -27,7 +27,7 @@ char	*ft_substr(char const *s, unsigned int start, size_t len);
 char	*get_next_line(int fd);
 int		ft_strchr(char *str, int c);
 char	*ft_strdup(const char *s1);
-size_t	ft_strlen(const char *str);
+int		ft_strlen(const char *str);
 char	*ft_strjoin(char *s1, char *s2);
 
 #endif
