+++ gnl/get_next_line_utils.c
@@ -6,7 +6,7 @@
 /*   By: mtacunan <mtacunan@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/12/07 22:04:45 by mtacunan          #+#    #+#             */
-/*   Updated: 2021/12/07 22:20:32 by mtacunan         ###   ########.fr       */
+/*   Updated: 2022/03/17 13:21:22 by mtacunan         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -49,7 +49,7 @@ char	*ft_substr(char const *s, unsigned int start, size_t len)
 	return (sol);
 }
 
-size_t	ft_strlen(const char *str)
+int	ft_strlen(const char *str)
 {
 	int	len;
